rename(ts/renderSvg): rename renderSvg to svgIcon to match new component
