+++ assets/src/helpers/svgIcon.tsx
